Crate teensyduino

Functions from the teensyduino framework, re-exported and wrapped in rusty, safe wrappers.

Re-exports

pub use gpio::Pin;

pub use gpio::PinMode;

pub use gpio::LED_BUILTIN;

pub use sound::Tone;

Modules

gpio
Tools for interfacing with GPIO on the uC

serial
Tooling for accessing the Serial and UART ports on the uC

sound
Sound functionality provided by the arduino framework

tempmon
Utilities for interfacing with the teensy 4's integrated temperature monitor

Functions

delay
Pauses the program for the amount of time (in milliseconds) specified as parameter. (There are 1000 milliseconds in a second.)

delay_microseconds
Pauses the program for the amount of time (in microseconds) specified by the parameter. There are a thousand microseconds in a millisecond and a million microseconds in a second.

delay_nanoseconds
Basically delay microseconds, but nanoseconds

micros
Returns the number of microseconds since the Arduino board began running the current program. This number will overflow (go back to zero), after approximately 70 minutes. On 16 MHz Arduino boards (e.g. Duemilanove and Nano), this function has a resolution of four microseconds (i.e. the value returned is always a multiple of four). On 8 MHz Arduino boards (e.g. the LilyPad), this function has a resolution of eight microseconds.

millis
Returns the number of milliseconds passed since the Arduino board began running the current program. This number will overflow (go back to zero), after approximately 50 days.