Function teensyduino::micros

pub fn micros() -> u32

Returns the number of microseconds since the Arduino board began running the current program. This number will overflow (go back to zero), after approximately 70 minutes. On 16 MHz Arduino boards (e.g. Duemilanove and Nano), this function has a resolution of four microseconds (i.e. the value returned is always a multiple of four). On 8 MHz Arduino boards (e.g. the LilyPad), this function has a resolution of eight microseconds.

Returns

Returns the number of microseconds since the Arduino board began running the current program.

Notes and Warnings

There are 1,000 microseconds in a millisecond and 1,000,000 microseconds in a second.

See Also

micros() - Arduino Reference