Module teensyduino::gpio

Tools for interfacing with GPIO on the uC

See Also

Digital Pins - Arduino Reference

Structs

Pin
A container for a pin number that has been verified to exist This method was chosen to limit the use of unsafe when dealing with pins

Enums

PinMode
The modes that a pin can be configured in

Constants

LED_BUILTIN
The builtin LED, normally on pin 13